requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
gitpython==3.1.40
chromadb==0.4.18
langchain==0.0.335
langchain-community==0.0.13
rich==13.6.0
slowapi==0.1.8
requests==2.31.0
numpy==1.26.2
matplotlib==3.8.2
networkx==3.2.1
pygments==2.16.1
pytest==7.4.3
httpx==0.25.1
